feat: add plugin architecture
