+++ lib/configValidator.js
@@ -5,7 +5,7 @@ exports.validateConfig = validateConfig;
 const allowedArchitectures = ["clean", "layered", "onion", "ddd"];
 const allowedApiTypes = ["rest", "graphql"];
 function validateConfig(config) {
-    var _a, _b, _c, _d, _e;
+    var _a, _b, _c, _d, _e, _f;
     const errors = [];
     if (!config.projectName) {
         errors.push("projectName is required");
@@ -43,5 +43,6 @@ function validateConfig(config) {
         enableHttpClients: (_c = config.enableHttpClients) !== null && _c !== void 0 ? _c : false,
         enableSwagger: (_d = config.enableSwagger) !== null && _d !== void 0 ? _d : false,
         enableCors: (_e = config.enableCors) !== null && _e !== void 0 ? _e : false,
+        plugins: (_f = config.plugins) !== null && _f !== void 0 ? _f : [],
     };
 }
