+++ plugins/logger.js
@@ -1,0 +1,3 @@
+module.exports.apply = async function(context) {
+  console.log(`logger plugin: generating ${context.config.projectName}`);
+};
